Split test up into several files
