+++ integration/testnet-clone.json
@@ -6,7 +6,7 @@
         }
     ],
     "contracts": {
-        "flow_cold_storage_proxy": "a72f9a62fb8ddbbf",
+        "flow_cold_storage_proxy": "e458b3a482e81bf5",
         "flow_fees": "912d5440f7e3769e",
         "flow_token": "7e60df042a9c0868",
         "fungible_token": "9a0766d93b6608b7"
@@ -15,7 +15,7 @@
     "disable_consensus_follower": true,
     "network": "testnet",
     "originators": [
-        "a72f9a62fb8ddbbf"
+        "e458b3a482e81bf5"
     ],
     "port": 8080,
     "spork_seal_tolerance": 600,
